+++ src/Controller/Admin/AppController.php
@@ -44,7 +44,7 @@ abstract class AppController extends BaseAppController
         $this->Authorization->authorize($this);
 
         //Sets paginate limit and maximum paginate limit
-        //See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
+        //See http://book.cakephp.org/4/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
         $this->paginate['limit'] = $this->paginate['maxLimit'] = getConfigOrFail('admin.records');
 
         $this->viewBuilder()->setClassName('MeCms.View/Admin/App');