+++ src/Model/Table/PostsTable.php
@@ -70,8 +70,7 @@ class PostsTable extends PostsAndPagesTables
             //Gets existing tags
             $existingTags = $this->Tags->getList()->toArray();
 
-            //For each tag, it searches if the tag already exists.
-            //If a tag exists in the database, it sets also the tag ID
+            //For each tag, it searches if the tag already exists. If a tag exists, it sets also the tag ID
             foreach ($tags as $k => $tag) {
                 $id = array_search($tag, $existingTags);
                 if ($id) {
@@ -116,7 +115,7 @@ class PostsTable extends PostsAndPagesTables
      * @param \MeCms\Model\Entity\Post $post Post entity. It must contain `id` and `Tags`
      * @param int $limit Limit of related posts
      * @param bool $images If `true`, gets only posts with images
-     * @return \Cake\Collection\CollectionInterface Collection of entities
+     * @return \Cake\Collection\CollectionInterface<\MeCms\Model\Entity\Post>
      * @throws \Tools\Exception\PropertyNotExistsException
      */
     public function getRelated(Post $post, int $limit = 5, bool $images = true): CollectionInterface
@@ -148,7 +147,7 @@ class PostsTable extends PostsAndPagesTables
                         continue;
                     }
 
-                    //Adds the current post to the related posts and its ID to the  IDs to be excluded for the next query
+                    //Adds the current post to the related posts and its ID to the IDs to be excluded for the next query
                     $related[] = $post;
                     $exclude[] = $post->get('id');
                 }