+++ src/TestSuite/EntityTestCase.php
@@ -57,9 +57,10 @@ abstract class EntityTestCase extends TestCase
         parent::setUp();
 
         if (empty($this->Entity) && $this->autoInitializeClass) {
-            /** @var class-string<\Cake\ORM\Entity> $className */
-            $className = $this->getOriginClassNameOrFail($this);
-            $this->Entity = new $className();
+            $entityClassName = $this->getOriginClassNameOrFail($this);
+            /** @var \Cake\ORM\Entity $Entity */
+            $Entity = new $entityClassName();
+            $this->Entity = $Entity;
         }
     }
 }