+++ tests/TestCase/Command/Install/CopyConfigCommandTest.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -28,7 +29,7 @@ class CopyConfigCommandTest extends TestCase
     use ConsoleIntegrationTestTrait;
 
     /**
-     * Test for `execute()` method
+     * @uses \MeCms\Command\Install\CopyConfigCommand::execute()
      * @test
      */
     public function testExecute(): void