+++ tests/TestCase/Command/Install/CreateVendorsLinksCommandTest.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -28,7 +29,7 @@ class CreateVendorsLinksCommandTest extends TestCase
     use ConsoleIntegrationTestTrait;
 
     /**
-     * Tests for `execute()` method
+     * @uses \MeTools\Command\Install\CreateVendorsLinksCommand::execute()
      * @test
      */
     public function testExecute(): void