+++ tests/TestCase/Model/Table/Traits/GetPreviewsFromTextTraitTest.php
@@ -1,4 +1,5 @@
 <?php
+/** @noinspection PhpUnhandledExceptionInspection */
 declare(strict_types=1);
 
 /**
@@ -16,7 +17,7 @@ declare(strict_types=1);
 namespace MeCms\Test\TestCase\Model\Table\Traits;
 
 use Cake\Collection\CollectionInterface;
-use Cake\ORM\Entity;
+use MeCms\Model\Table\PostsTable;
 use MeCms\TestSuite\TestCase;
 use MeTools\Utility\Youtube;
 
@@ -28,7 +29,7 @@ class GetPreviewsFromTextTraitTest extends TestCase
     /**
      * @var \MeCms\Model\Table\PostsTable&\PHPUnit\Framework\MockObject\MockObject
      */
-    protected $Posts;
+    protected PostsTable $Posts;
 
     /**
      * Called before every test method
@@ -38,11 +39,15 @@ class GetPreviewsFromTextTraitTest extends TestCase
     {
         parent::setUp();
 
-        $this->Posts = $this->Posts ?: $this->getTable('MeCms.Posts');
+        if (empty($this->Posts)) {
+            /** @var \MeCms\Model\Table\PostsTable&\PHPUnit\Framework\MockObject\MockObject $Posts */
+            $Posts = $this->getTable('MeCms.Posts');
+            $this->Posts = $Posts;
+        }
     }
 
     /**
-     * Test for `extractImages()` method
+     * @uses \MeCms\Model\Table\Traits\GetPreviewsFromTextTrait::extractImages()
      * @test
      */
     public function testExtractImages(): void
@@ -132,17 +137,7 @@ class GetPreviewsFromTextTraitTest extends TestCase
     }
 
     /**
-     * Test for `getPreviewSize()` method
-     * @test
-     */
-    public function testGetPreviewSize(): void
-    {
-        $result = $this->invokeMethod($this->Posts, 'getPreviewSize', [WWW_ROOT . 'img' . DS . 'image.jpg']);
-        $this->assertEquals([400, 400], $result);
-    }
-
-    /**
-     * Test for `getPreviews()` method
+     * @uses \MeCms\Model\Table\Traits\GetPreviewsFromTextTrait::getPreviews()
      * @test
      */
     public function testGetPreviews(): void
@@ -167,27 +162,41 @@ class GetPreviewsFromTextTraitTest extends TestCase
         $result = $Posts->getPreviews('<img src=\'' . WWW_ROOT . 'img' . DS . 'noExisting.jpg\' />');
         $this->assertTrue($result->isEmpty());
 
-        $result = $Posts->getPreviews('<img src=\'http://example.com/image.jpg\' />');
+        $result = $Posts->getPreviews('<img src=\'https://example.com/image.jpg\' />');
         $this->assertInstanceOf(CollectionInterface::class, $result);
         $this->assertCount(1, $result);
-        $this->assertContainsOnlyInstancesOf(Entity::class, $result);
-        $this->assertEquals('http://example.com/image.jpg', $result->first()->get('url'));
-        $this->assertEquals(400, $result->first()->get('width'));
-        $this->assertEquals(300, $result->first()->get('height'));
+        $this->assertEquals([
+            'url' => 'https://example.com/image.jpg',
+            'width' => 400,
+            'height' => 300,
+        ], $result->first());
 
         foreach (['image.jpg', WWW_ROOT . 'img' . DS . 'image.jpg'] as $image) {
             $result = $Posts->getPreviews('<img src=\'' . $image . '\' />');
             $this->assertCount(1, $result);
-            $this->assertMatchesRegularExpression('/^http:\/\/localhost\/thumb\/[A-z0-9]+$/', $result->first()->get('url'));
-            $this->assertEquals(400, $result->first()->get('width'));
-            $this->assertEquals(300, $result->first()->get('height'));
+            $first = $result->first();
+            $this->assertMatchesRegularExpression('/^http:\/\/localhost\/thumb\/[A-z0-9]+$/', $first['url']);
+            $this->assertSame(400, $first['width']);
+            $this->assertSame(300, $first['height']);
         }
 
         $youtubeId = '6z4KK7RWjmk';
         $result = $Posts->getPreviews('[youtube]' . $youtubeId . '[/youtube]');
         $this->assertCount(1, $result);
-        $this->assertEquals(Youtube::getPreview($youtubeId), $result->first()->get('url'));
-        $this->assertEquals(400, $result->first()->get('width'));
-        $this->assertEquals(300, $result->first()->get('height'));
+        $this->assertEquals([
+            'url' => Youtube::getPreview($youtubeId),
+            'width' => 400,
+            'height' => 300,
+        ], $result->first());
+    }
+
+    /**
+     * @uses \MeCms\Model\Table\Traits\GetPreviewsFromTextTrait::getPreviewSize()
+     * @test
+     */
+    public function testGetPreviewSize(): void
+    {
+        $result = $this->invokeMethod($this->Posts, 'getPreviewSize', [WWW_ROOT . 'img' . DS . 'image.jpg']);
+        $this->assertEquals([400, 400], $result);
     }
 }